Moved the analysis function into separate file path_species_anlysis.py (unpolluting discretize_path).
