+++ tropical/discretize_path.py
@@ -385,88 +385,3 @@ def natural_sort(l):
     convert = lambda text: int(text) if text.isdigit() else text.lower()
     alphanum_key = lambda key: [convert(c) for c in re.split('([0-9]+)', key)]
     return sorted(l, key=alphanum_key)
-
-def get_path_descendants(path):
-    """ Get the set of descendants for a tree like path dict.
-    """
-
-    importer = DictImporter()
-    root = importer.import_(path)
-    descendants = set([descendant_node.name for descendant_node in root.descendants])
-    return descendants
-
-def global_conserved_species_analysis(paths, path_signatures, model):
-    """
-    Computes the fraction of dominant paths a species is in. It is taken over
-        all simulations and all timepoints.
-    Parameters
-    ----------
-    paths: dict
-        Nested tree structure dict of paths as returned from
-            DomPath.get_path_signatures()
-    path_signatures: pandas.DataFrame
-        The dominant path signatures for each simulation (across all
-            timepoints).
-    model: pysb.Model
-        The model that is being used.
-
-    Returns
-    -------
-    A list of tuples with the species codename
-        (i.e. 's' + str( model.species_index)) and the fraction of dominant
-        paths that species was in.
-    """
-
-    def convert_names(list_o_tuple):
-        new_list_o_tuple = []
-        for i, item in enumerate(list_o_tuple):
-            sname = item[0]
-            node_idx = list(find_numbers(sname))[0]
-            node_sp = model.species[node_idx]
-            node_name = parse_name(node_sp)
-            new_list_o_tuple.append((node_name, item[1]))
-
-        return new_list_o_tuple
-    generate_equations(model)
-    species_all = model.species
-    #print(species_all)
-    n_species_all = len(species_all)
-    spec_dict = dict()
-    spec_counts = np.array([0.0] * n_species_all)
-    #species_all_snames = []
-    for i, species in enumerate(species_all):
-        sname = "s{}".format(i)
-        spec_dict[sname] = {'name': species, 'index': i}
-
-    path_species = dict()
-    for i, key in enumerate(paths.keys()):
-        path = paths[key]
-        descendants = get_path_descendants(path)
-        #add the root node to the set of species for the path
-        descendants.add(path['name'])
-        #print(descendants)
-        path_species[i] = descendants
-    path_signatures_np = path_signatures.values
-    n_sims = path_signatures_np.shape[0]
-    n_tp = path_signatures_np.shape[1]
-    #print(n_sims, n_tp)
-    #quit()
-    n_tot = 0.0
-    for i in range(n_sims):
-        for j in range(n_tp):
-            n_tot += 1.0
-            dom_path_id = path_signatures_np[i][j]
-            #print(dom_path_id)
-            for descendant in path_species[dom_path_id]:
-            #    print(descendant)
-                d_id = spec_dict[descendant]['index']
-                spec_counts[d_id] += 1.0
-    #print(n_tot)
-    spec_fracs = spec_counts / n_tot
-    #quit()
-    spec_frac_dict = dict()
-    for spec in spec_dict.keys():
-        spec_frac_dict[spec] = spec_fracs[spec_dict[spec]['index']]
-    sorted_by_value = sorted(spec_frac_dict.items(), key=lambda kv: -kv[1])
-
-    return convert_names(sorted_by_value)
